@@ -6,8 +6,14 @@ import java.util.List;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.Resource;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.dao.DataAccessException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -17,11 +23,17 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.marcorh96.springboot.rest.ecommerce.app.models.document.Product;
 import com.marcorh96.springboot.rest.ecommerce.app.models.services.IProductService;
+import com.marcorh96.springboot.rest.ecommerce.app.models.services.file.IUploadFileService;
 import com.mongodb.DuplicateKeyException;
+
+import java.io.IOException;
+import java.net.MalformedURLException;
 
 @RestController
 @RequestMapping("/api")
@@ -30,10 +42,20 @@ public class ProductRestController {
     @Autowired
     private IProductService productService;
 
+    @Autowired
+    @Qualifier("upload-products")
+    private IUploadFileService uploadFileService;
+
     @GetMapping("/products")
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_USER')")
-    public List<Product> show() {
+    public List<Product> showProducts() {
         return productService.findAll();
+    }
+
+    @GetMapping("/products/page/{page}")
+    public Page<Product> showProducts(@PathVariable Integer page) {
+        Pageable pageable = PageRequest.of(page, 2);
+        return productService.findAll(pageable);
     }
 
     @GetMapping("/products/{id}")
@@ -70,6 +92,9 @@ public class ProductRestController {
     public ResponseEntity<?> deleteProduct(@PathVariable String id) {
         Map<String, Object> response = new HashMap<>();
         try {
+            Product product = productService.findById(id);
+            String photoPastName = product.getPhoto();
+            uploadFileService.delete(photoPastName);
             productService.delete(id);
         } catch (DataAccessException e) {
             response.put("message", "Data Base Exception!");
@@ -102,5 +127,44 @@ public class ProductRestController {
         response.put("message", "Product has been updated successfully!");
         return new ResponseEntity<Map<String, Object>>(response, HttpStatus.CREATED);
 
+    }
+
+    @PostMapping("/products/upload")
+    public ResponseEntity<?> upload(@RequestParam("file") MultipartFile file, @RequestParam("id") String id) {
+        Map<String, Object> response = new HashMap<>();
+        Product product = productService.findById(id);
+        if (!file.isEmpty()) {
+            String nombreArchivo = null;
+            try {
+                nombreArchivo = uploadFileService.copy(file);
+            } catch (IOException e) {
+                response.put("message", "Error uploading image" + nombreArchivo);
+                response.put("error", e.getMessage().concat(": ").concat(e.getCause().getMessage()));
+                return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+            String photoPastName = product.getPhoto();
+
+            uploadFileService.delete(photoPastName);
+            product.setPhoto(nombreArchivo);
+            productService.save(product);
+            response.put("product", product);
+            response.put("message", "Image has been uploaded successfully: " + nombreArchivo);
+        }
+        return new ResponseEntity<Map<String, Object>>(response, HttpStatus.CREATED);
+    }
+
+    @GetMapping("/products/upload/img/{photoName:.+}")
+    public ResponseEntity<Resource> viewPhoto(@PathVariable String photoName) {
+
+        Resource resource = null;
+        try {
+            resource = uploadFileService.load(photoName);
+        } catch (MalformedURLException e) {
+            e.printStackTrace();
+        }
+        HttpHeaders headers = new HttpHeaders();
+        headers.add(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + resource.getFilename() + "\"");
+
+        return new ResponseEntity<Resource>(resource, headers, HttpStatus.OK);
     }
 }
